feat(app): implement check-in flows, inventory refactor, and access code generation
- **Check-ins:** Created `DailyCheckinActivity` and `DailyCheckinActivityParent` with `CheckinService` for logging symptoms, triggers, feelings, and peak flow data to Firestore.
- **Inventory:** Refactored `InventoryItem` and `InventoryRepository` to strictly support "Rescue" and "Controller" types. Updated `ChildrenAdapter` to manage canister details and calculate asthma zone thresholds (Green/Yellow/Red) based on personal best.
- **Access Code:** Added `PacGenerator` utility and a dialog in `ParentHomeActivity` for generating and viewing Parent Access Codes. Updated `InputValidator` regex to match the new format.
- **Providers:** Enhanced `ProvidersAdapter` and `ActivitySettingsProvider` to include UI flows for sending history and real-time reports, including symptom and trigger selection dialogs.
- **Refactor:** Removed dummy test data from settings activities, moved `InventoryItem` to the `childcollections` package, and registered new activities in `AndroidManifest.xml`.

+++ app/src/main/java/com/example/smartair/models/childcollections/Checkin.java
@@ -2,6 +2,8 @@ package com.example.smartair.models.childcollections;
 
 import com.google.firebase.Timestamp;
 import java.util.List;
+import java.util.HashMap; // <-- New Import Needed
+import java.util.Map;
 
 /**
  * Purpose: Represents a daily check-in document stored in
@@ -16,7 +18,7 @@ public class Checkin {
     private String feeling;
     private List<String> symptoms;
     private List<String> triggers;
-    private Long peakFlow;          // may be null
+    private int peakFlow;          // may be null
     private EntryAuthor entryAuthor;
 
     /** Required for Firebase deserialization. */
@@ -27,7 +29,7 @@ public class Checkin {
                    String feeling,
                    List<String> symptoms,
                    List<String> triggers,
-                   Long peakFlow,
+                   int peakFlow,
                    EntryAuthor entryAuthor) {
         this.createdAt = createdAt;
         this.feeling = feeling;
@@ -69,11 +71,11 @@ public class Checkin {
         this.triggers = triggers;
     }
 
-    public Long getPeakFlow() {
+    public int getPeakFlow() {
         return peakFlow;
     }
 
-    public void setPeakFlow(Long peakFlow) {
+    public void setPeakFlow(int peakFlow) {
         this.peakFlow = peakFlow;
     }
 
@@ -84,4 +86,6 @@ public class Checkin {
     public void setEntryAuthor(EntryAuthor entryAuthor) {
         this.entryAuthor = entryAuthor;
     }
+
+
 }
